requirements.txt
flask
pillow
python-dotenv
google-generativeai
loguru